app/requirements.txt
requests
boto3
qrcode
pillow